added GetWithParameter method
