+++ shopify/shopify.go
@@ -2,7 +2,6 @@
 package shopify
 
 import (
-	"encoding/json"
 	"fmt"
 
 	"github.com/parnurzeal/gorequest"
@@ -28,30 +27,12 @@ func New(store, apiKey, pass string) Shopify {
 	return Shopify{store: store, apiKey: apiKey, pass: pass}
 }
 
-// Creates target URL for making a Shopify Request to a given endpoint
-func (shopify *Shopify) createTargetURL(endpoint string) string {
-	result := fmt.Sprintf("https://%s:%s@%s%s/%s.json", shopify.apiKey, shopify.pass, shopify.store, domain, endpoint)
-	return result
-}
-
-// getJSONBytesFromMap Extracts Json Bytes from map[string]interface
-func getJSONBytesFromMap(data map[string]interface{}) ([]byte, error) {
-	jsonData, err := json.Marshal(data)
-	if err != nil {
-		fmt.Println("Invalid data object, can't parse to json:")
-		fmt.Println("Error:", err)
-		fmt.Println("Data:", data)
-		return nil, err
-	}
-	return jsonData, nil
-}
-
 // Request Creates a new Request to Shopify and returns the response as a map[string]interface{}.
 // method: GET/POST/PUT - string
 // url: target endpoint like "products" - string
 // data: content to be sent with the request
 // Usage: shopify.request("GET","products",nil)
-func (shopify *Shopify) Request(method, endpoint string, data map[string]interface{}) ([]byte, []error) {
+func (shopify *Shopify) Request(method, endpoint string, data interface{}) ([]byte, []error) {
 	jsonData, _ := getJSONBytesFromMap(data)
 	targetURL := shopify.createTargetURL(endpoint)
 
@@ -72,7 +53,12 @@ func (shopify *Shopify) Request(method, endpoint string, data map[string]interfa
 // shopify.Get("products/5.json")
 // shopify.Get("products/5/variants.json")
 func (shopify *Shopify) Get(endpoint string) ([]byte, []error) {
-	targetURL := shopify.createTargetURL(endpoint)
+	return shopify.GetWithParameters(endpoint, nil)
+}
+
+// GetWithParameters Makes a GET request to shopify with the given endpoint and given parameters
+func (shopify *Shopify) GetWithParameters(endpoint string, parameters map[string]string) ([]byte, []error) {
+	targetURL := shopify.createTargetURLWithParameters(endpoint, parameters)
 	request := gorequest.New()
 	_, body, errs := request.Get(targetURL).End()
 
@@ -81,7 +67,7 @@ func (shopify *Shopify) Get(endpoint string) ([]byte, []error) {
 
 // Post Makes a POST request to shopify with the given endpoint and data.
 // Usage: shopify.Post("products", map[string]interface{} = product data map)
-func (shopify *Shopify) Post(endpoint string, data map[string]interface{}) ([]byte, []error) {
+func (shopify *Shopify) Post(endpoint string, data interface{}) ([]byte, []error) {
 	targetURL := shopify.createTargetURL(endpoint)
 	jsonData, err := getJSONBytesFromMap(data)
 	if err != nil {
@@ -100,7 +86,7 @@ func (shopify *Shopify) Post(endpoint string, data map[string]interface{}) ([]by
 
 // Put Makes a PUT request to shopify with the given endpoint and data.
 // Usage: shopify.Put("products", map[string]interface{} = product data map)
-func (shopify *Shopify) Put(endpoint string, data map[string]interface{}) ([]byte, []error) {
+func (shopify *Shopify) Put(endpoint string, data interface{}) ([]byte, []error) {
 	targetURL := shopify.createTargetURL(endpoint)
 	jsonData, err := getJSONBytesFromMap(data)
 	if err != nil {
@@ -126,4 +112,21 @@ func (shopify *Shopify) Delete(endpoint string) ([]byte, []error) {
 	_, body, errs := request.Delete(targetURL).End()
 
 	return []byte(body), errs
+}
+
+// Creates target URL for making a Shopify Request to a given endpoint
+func (shopify *Shopify) createTargetURL(endpoint string) string {
+	return shopify.createTargetURLWithParameters(endpoint, nil)
+}
+
+// Creates target URL for making a Shopify Request to a given endpoint with the given parameters
+func (shopify *Shopify) createTargetURLWithParameters(endpoint string, parameters map[string]string) string {
+	var parametersString = ""
+	if parameters != nil && len(parameters) > 0 {
+		parametersString = "?"
+		for k := range parameters {
+			parametersString = fmt.Sprintf("%v%v=%v&", parametersString, k, parameters[k])
+		}
+	}
+	return fmt.Sprintf("https://%s:%s@%s%s/%s.json%s", shopify.apiKey, shopify.pass, shopify.store, domain, endpoint, parametersString)
 }
